Oops, almost forgot errors

+++ exercise.wwwapi/Endpoints/ProductEndpoints.cs
@@ -1,7 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using exercise.wwwapi.DTO;
+﻿using exercise.wwwapi.DTO;
 using exercise.wwwapi.Model;
 using exercise.wwwapi.Repository;
+using Microsoft.AspNetCore.Mvc;
+using System.Xml.Linq;
 
 namespace exercise.wwwapi.Endpoints
 {
@@ -20,8 +21,16 @@ namespace exercise.wwwapi.Endpoints
         }
 
         [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public static async Task<IResult> CreateProduct(IProductRepository repository, string name, string category, int price)
         {
+            var productsList = await repository.GetAllProducts();
+
+            if (!(price is int) || productsList.Any(x => x.Name.Equals(name)))
+            {
+                return TypedResults.BadRequest("Price must be an integer, something else was provided. / Product with provided name already exists.");
+            }
+
             var newProduct = new Product
             {
                 Name = name,
@@ -35,22 +44,52 @@ namespace exercise.wwwapi.Endpoints
         }
 
         [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public static async Task<IResult> GetAllProducts(IProductRepository repository)
         {
             var results = await repository.GetAllProducts();
+
+            if (!results.Any())
+            {
+                return TypedResults.NotFound("No products of provided category were found");
+            }
+
             return TypedResults.Ok(results);
         }
 
         [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public static async Task<IResult> GetProductsById(IProductRepository repository, int id)
         {
             var result = await repository.GetProductById(id);
+
+            if (result == null)
+            {
+                return TypedResults.NotFound("Product not found");
+            }
+
             return TypedResults.Ok(result);
         }
 
         [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public static async Task<IResult> UpdateProduct(IProductRepository repository, string name, int id, string category, int price)
         {
+            var result = await repository.GetProductById(id);
+
+            if (result == null)
+            {
+                return TypedResults.NotFound("Product not found");
+            }
+
+            var productsList = await repository.GetAllProducts();
+
+            if (!(price is int) || productsList.Any(x => x.Name.Equals(name)))
+            {
+                return TypedResults.BadRequest("Price must be an integer, something else was provided. / Product with provided name already exists.");
+            }
+
             var newProduct = new Product
             {
                 Id = id,
@@ -64,9 +103,19 @@ namespace exercise.wwwapi.Endpoints
         }
 
         [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public static async Task<IResult> DeleteProduct(IProductRepository repository, int id)
         {
+
+            var result = await repository.GetProductById(id);
+
+            if (result == null)
+            {
+                return TypedResults.NotFound("Product not found");
+            }
+
             var deleted = await repository.DeleteProduct(id);
+
             return TypedResults.Ok(deleted);
         }
     }
